# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CraftedBy API",
        "version": "0.1.0"
    },
    "paths": {
        "/business": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get all businesses",
                "operationId": "d5565e76960829e56bd5e12b42e867a5",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Update business",
                "operationId": "6527223db8f5fa3c3f36ac7a29dd2d90",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete business",
                "operationId": "c3ded99ba6595f4c45124f37c983f5ef",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/business/{id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get business",
                "operationId": "2fc11a0dcbe9f8be3cac3928a162356c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "color",
                        "in": "query",
                        "description": "Add color filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "material",
                        "in": "query",
                        "description": "Add material filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "style",
                        "in": "query",
                        "description": "Add style filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Add category filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Add size filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business",
                        "in": "query",
                        "description": "Add business filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get one product",
                "operationId": "b5099368868dbe7900acd129345eb148",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "operationId": "2ce632b9d1c0c445557821d09add7ec2",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "operationId": "76bb7dc4ab3fc3233d5b33ef58e95d7d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "7208c29037df2b8bb02499ba39ed8a78",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get one user",
                "operationId": "6946bbab7898f59a09e3c8e529a38907",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "operationId": "05028d02ea3e9a99abd6260cfe05b179",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "b3d28da5dc9de7fc1bd425f93b54ff7b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "image_path": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "active": {
                        "type": "tinyint"
                    },
                    "description": {
                        "type": "string"
                    },
                    "color_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "material_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "style_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "size_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "business_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            }
        }
    }
}